LayoutTests/js/dom/modules/module-incorrect-relative-specifier-expected.txt

CONSOLE MESSAGE: TypeError: Module specifier does not start with "/", "./", or "../".
CONSOLE MESSAGE: TypeError: Module specifier does not start with "/", "./", or "../".
CONSOLE MESSAGE: TypeError: Module specifier does not start with "/", "./", or "../".
Test script.onerror will be fired when the incorrect relative specifier is specified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
0
1
2
PASS scripts.has(document.getElementById("s0")) is true
PASS scripts.has(document.getElementById("s1")) is true
PASS scripts.has(document.getElementById("s2")) is true
PASS successfullyParsed is true

TEST COMPLETE
